requirements.txt
strands-agents
strands-agents-tools
mcp
anthropic
requests
python-dotenv
